adds a method to display image planes on a rop
adds an additional method that would display the existing image planes
on a rop. the reasoning being, it is currently a little hard to parse
the houdini content on a rop as it is kind of long. The long term plan
is to give the user the optionality to enter custom image plane names in
the same format to be mass added by the tool.